requirements.txt
# Core Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
python-multipart==0.0.6

# HTTP Client
requests==2.31.0
httpx==0.25.2
aiohttp==3.9.1

# LLM Providers
openai==1.3.0
anthropic==0.8.0
langchain==0.1.0
langchain-openai==0.0.2
ollama==0.1.3

# Background Tasks / Queue
celery[redis]==5.3.4
redis==5.0.1

# Scientific Computing
numpy==1.24.3
scipy==1.11.4
pandas==2.1.3
matplotlib==3.8.2

# Configuration / Serialization
pyyaml==6.0.1
toml==0.10.2

# CLI
click==8.1.7
rich==13.7.0

# Process Management
psutil==5.9.7

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Code Quality
black==23.12.0
flake8==6.1.0
mypy==1.7.1
isort==5.13.2

# Type Stubs
types-requests==2.31.0.10
types-PyYAML==6.0.12.12
